Add axis repositioning across x, y, and, z axises
Axis Repositioning is activated by offhand gripping and holding
the trigger (c_repositionButtonID) and saved when released.
Once in repositioning mode and holding trigger, the axis is cycled
by clicking on the A button (c_defaultPositionButtonID). The distance
on the axis is tied to the offhand button moves. Currently, the offhand
moves only on the X-axis, but the weapon will preview the final location