Add in angular maximum to dynamic grip
Currently just a dirty fix for melee weapons being rotated by 70 degrees with dynamic grip on. Could be expanded as a comfort option if given an INI setting. Should be replaced or supplemented with a better method for detecting melee weapons, which need to remain at their original rotation to match with their effective hitbox.