Fix blurry of 3D objects rendering in specific on screen views (screenshots below): 1. PipBoy in "In Front" config 2. Transfer of items UI 3. All crafting UIs
repro: open one of the UIs and move your head. the 3D object looks really bad.

I bisect the bug to v67 release and then to 6b21c428 change of `fixMelee` function in `Skeleton.cpp`. The reason it got worse there is because of "isMelee" check was passing for no equipped weapon. Probably because fists is a melee weapon and you can have grenades equipped.

But the reason for the bad UI is actually in `(*g_player)->Update(0.0f);` line. Removing it fixes the visual issue without any negative effects that I can tell.

Additional changes: I refactored the code to do less as we can return early if no weapon is drawn or if the weapon is not melee.