Improve performance of cullGeometry by reducing the amount of times we iterate over all the player geometries to only once a second and not every frame.
It reduces the "cullPlayerGeometry" time from ~0.05ms to ~0.0002ms (for 89 frames, if framerate is 90).
May not sounds as much but total of this mod update frame is 0.25ms making it 20% of the time!
And god dammit the game is slow enough not to waste more time.

And refactor the code into it's own handler