Fix right hand incorrect positioning in left handed mode when a throwable weapon is equipped.
Reason: equipping throwable weapon adds another "WeaponLeft" node to the skeleton left arm. this result in what was supposed to be right arm node (for left-handed) to be left arm node. Which results in very incorrect right-hand position.
The extra node stays in the skeleton even when unequipping the throwable. That's why only restarting fixed the issue.

I've used `_playerNodes->WeaponLeftNode` instead of searching for the node by name as it looks to do the correct thing. The other option is to search for it under the right arm like so:
```
auto fpSkelly = (*g_player)->firstPersonSkeleton->GetAsNiNode();
auto leftWeaponFindRoot = g_config->leftHandedMode ? getNode( "RArm_Collarbone", fpSkelly) : fpSkelly;
NiNode* leftWeapon = getNode("WeaponLeft", leftWeaponFindRoot);
```
in case there is a problem we can change to this code.