Refactor and improve smooth movement - Fix to smooth movement when the game thinks the player is in the air - The solution is disabling veritcal smooth movement when in the air but leaving horizontal smoothing   - Fixes jump landing jitter   - Fixes jitter in area where there is a small gap in the ground like the exist platform of vault 111
I tried to detect if the "in the air" is only for a few frames but that results in jump jitter.
It means there is a small jitter when the game thinks the player is in the air for 5-15 frames but it's not too bad and most players won't notice it.