Move to CommonLibF4 - fix row/column major multiplication in WeaponPositionAdjuster   - To prevent breaking everything fix the multiplication in one class by using `matrixVecMultTempFix` and `updateTransformsFixed`
As row/column major multiplication is basically a transpose done before doing the regular multiplication we need to remove existing transpose when moving to local space and add transpose when moving to world space. The we can use the correct matrix-vector and matrix-matrix multiplication removing a lot of custom code that had transpose embedded in it.

Fix steps:

For every location that does matrix-vector multiplication:
  - if moving to local space:
    1. remove transpose
    2. use `matrixVecMultTempFix`
  - else
    1. add transpose
    2. use `matrixVecMultTempFix`

For every location that does matrix-matrix multiplication (multiply43Left)
  1. Use NiMatrix3 multiplication directly
    - Use `.getMatrix3()` to get matrix from quaternion
    - use the correct right/left multiplication order by what method was used `multiply43Left` or `multiply43Right`

Update calls of `updateTransforms` to `updateTransformsFixed` that uses the correct and direct matrix multiplication