Fix bug where main menu is not opened after exiting power armor
There is a bug that when existing Power Armor you can't open the main menu.
But it only happens if the Pipboy is on-wrist setting (others are fine).
A fix is to open and close the Pipboy.
The hack here is to open the Pipboy in the ctor and close it ASAP in the first frame.